+++ packages/create-cloudflare/e2e-tests/frameworks/framework-test-config-experimental.ts
@@ -2,37 +2,6 @@ import { keys, LONG_TIMEOUT } from "../helpers";
 
 export default function getFrameworkTestConfigExperimental() {
 	return {
-		next: {
-			testCommitMessage: true,
-			flags: [
-				"--ts",
-				"--tailwind",
-				"--eslint",
-				"--app",
-				"--import-alias",
-				"@/*",
-				"--src-dir",
-			],
-			verifyBuildCfTypes: {
-				outputFile: "cloudflare-env.d.ts",
-				envInterfaceName: "CloudflareEnv",
-			},
-			verifyPreview: {
-				route: "/test",
-				expectedText: "Create Next App",
-			},
-			verifyDeploy: {
-				route: "/",
-				expectedText: "Create Next App",
-			},
-			// see https://github.com/cloudflare/next-on-pages/blob/main/packages/next-on-pages/docs/supported.md#operating-systems
-			unsupportedOSs: ["win32"],
-			unsupportedPms: [
-				// bun and yarn are failing in CI
-				"bun",
-				"yarn",
-			],
-		},
 		solid: {
 			promptHandlers: [
 				{
