+++ packages/create-cloudflare/e2e-tests/frameworks/framework-test-config.ts
@@ -360,7 +360,8 @@ export default function getFrameworkTestConfig(pm: string) {
 			},
 			flags: ["--typescript", "--no-install", "--no-git-init"],
 		},
-		next: {
+		"next:pages": {
+			argv: ["--platform", "pages"],
 			promptHandlers: [
 				{
 					matcher: /Do you want to use the next-on-pages eslint-plugin\?/,
@@ -392,6 +393,38 @@ export default function getFrameworkTestConfig(pm: string) {
 				"--turbopack",
 				"--import-alias",
 				"@/*",
+			],
+		},
+		"next:workers": {
+			argv: ["--platform", "workers"],
+			testCommitMessage: true,
+			flags: [
+				"--ts",
+				"--tailwind",
+				"--eslint",
+				"--app",
+				"--import-alias",
+				"@/*",
+				"--src-dir",
+			],
+			verifyBuildCfTypes: {
+				outputFile: "cloudflare-env.d.ts",
+				envInterfaceName: "CloudflareEnv",
+			},
+			verifyPreview: {
+				route: "/test",
+				expectedText: "Create Next App",
+			},
+			verifyDeploy: {
+				route: "/",
+				expectedText: "Create Next App",
+			},
+			// see https://github.com/cloudflare/next-on-pages/blob/main/packages/next-on-pages/docs/supported.md#operating-systems
+			unsupportedOSs: ["win32"],
+			unsupportedPms: [
+				// bun and yarn are failing in CI
+				"bun",
+				"yarn",
 			],
 		},
 		"nuxt:pages": {
