+++ packages/create-cloudflare/templates/next/pages/README.md
