+++ packages/create-cloudflare/templates/next/pages/app/js/app/api/hello/route.js
