+++ packages/create-cloudflare/templates/next/pages/app/js/app/not-found.js
