+++ packages/create-cloudflare/templates/next/pages/app/ts/app/api/hello/route.ts
