+++ packages/create-cloudflare/templates/next/pages/app/ts/app/not-found.tsx
