+++ packages/create-cloudflare/templates/next/pages/env.d.ts
