+++ packages/create-cloudflare/templates/next/pages/pages/js/pages/api/hello.js
