+++ packages/create-cloudflare/templates/next/pages/pages/ts/pages/api/hello.ts
