+++ packages/create-cloudflare/templates/next/pages/wrangler.jsonc
