+++ packages/create-cloudflare/templates/next/workers/templates/.dev.vars
@@ -1,0 +1,2 @@
+# Load .env.development* files when running `wrangler dev`
+NEXTJS_ENV=development
