+++ packages/create-cloudflare/templates/next/workers/templates/cloudflare-env.d.ts
@@ -1,0 +1,5 @@
+// Generated by Wrangler
+// by running `wrangler types --env-interface CloudflareEnv cloudflare-env.d.ts`
+
+interface CloudflareEnv {
+}
