+++ packages/create-cloudflare/templates/next/workers/templates/open-next.config.ts
@@ -1,0 +1,6 @@
+import { defineCloudflareConfig } from "@opennextjs/cloudflare";
+import kvIncrementalCache from "@opennextjs/cloudflare/kv-cache";
+
+export default defineCloudflareConfig({
+  incrementalCache: kvIncrementalCache,
+});
