+++ packages/create-cloudflare/templates/next/workers/templates/wrangler.jsonc
@@ -1,0 +1,12 @@
+{
+  "name": "<app-name>",
+  "main": ".open-next/worker.js",
+  "compatibility_date": "2024-09-26",
+  "compatibility_flags": [
+    "nodejs_compat"
+  ],
+  "assets": {
+    "binding": "ASSETS",
+    "directory": ".open-next/assets"
+  }
+}
